python-server/requirement.txt
grpcio
grpcio-tools
spacy==3.7.3
grpcio-reflection
numpy
openai
sentence_transformers
transformers
